+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - In `src/music.f90`, a subroutine `write_broken_chord()`, using an array containing the intervals to play. For the moment, each note has the same duration.
 - `ROADMAP.md`
 - An `example/README.md` file presenting each example.
+- `example/la_folia.f90`: variations on [La Folia](https://en.wikipedia.org/wiki/Folia), demonstrating the use of the subroutine `write_broken_chord()`.
 - A `logo/` directory.
 
 ### Changed